Introduce `nano::default_database_backend ()` getter
