+++ nano/node/nodeconfig.cpp
@@ -742,53 +742,3 @@ std::optional<unsigned> nano::node_config::env_io_threads ()
 	}();
 	return value;
 }
-
-std::optional<nano::database_backend> nano::node_config::env_database_backend ()
-{
-	static auto const backend = [] () -> std::optional<nano::database_backend> {
-		if (auto value = nano::env::get<std::string> ("NANO_BACKEND"))
-		{
-			auto backend = parse_database_backend (*value);
-			if (backend.has_value ())
-			{
-				std::cerr << "Default database backend overridden by NANO_BACKEND environment variable: " << to_string (*backend) << std::endl;
-			}
-			else
-			{
-				std::cerr << "Unknown database backend in NANO_BACKEND environment variable: " << *value << std::endl;
-			}
-			return backend;
-		}
-		return std::nullopt;
-	}();
-	return backend;
-}
-
-/*
- * database_backend
- */
-
-std::string nano::to_string (nano::database_backend const value)
-{
-	switch (value)
-	{
-		case nano::database_backend::lmdb:
-			return "lmdb";
-		case nano::database_backend::rocksdb:
-			return "rocksdb";
-	}
-	release_assert (false);
-}
-
-std::optional<nano::database_backend> nano::parse_database_backend (std::string const & value)
-{
-	if (value == "lmdb")
-	{
-		return nano::database_backend::lmdb;
-	}
-	if (value == "rocksdb")
-	{
-		return nano::database_backend::rocksdb;
-	}
-	return {};
-}
